Refactor data formats
In TBC the most annoying part of this project was manually entering data for each item as well as maintaining each character's data. In WotLK, since each character will have two specs and there are more opportunities to get gear, this will become an even greater annoyance.

To avoid having to set this data in code and to support a visual UI for managing this data the formats of the character data and the item databases have been refactored. Eventually character data and possibly the item database will be stored in localstorage and will be managed through the tracker directly.